Cleanup test/e2e/common directory - part2
